add KETREW_CONFIGURATION to configuration.env
throughout my humble trials and errors in working with coclo/epidisco,
I found myself forgetting to set this up and almost always adding appending
this to the `configuration.env`. Assuming that *most* people will be following
this tutorial and will only use the simplest setup, it might make sense to
make this default for convenience.