Filter spliced reads for IndelRealignment
We split RNA into two BAMs after realignment and some post-processing;
one with all the reads that span a splice junction (i.e. those with
CIGAR strings with an N in them). We IndelRealign unspliced reads, and
then merge the spliced reads back into the final BAM. Do we this because
the IndelRealigner will drop all reads in a pileup after a spliced read
is found, which can end up dropping the vast majority of our reads. We
want to continue running IndelRealignment, though, in order to retain as
much RNA support for indels as possible.